Remove IDocument and its implementations, part 1.
* Move FnDocument.get_ActiveMarkings to DocumentTraverser.

* Replace FnDocument.getContent with accessContentStream (a FileNet
  API convenience method that does exactly what getContent was doing).
  Add error handling for unsupported content elements (see b/6513927).

* Move the getProperyNames and getProperty methods to a separate
  IDocumentProperties interface. Make getPropertyStringValue and
  getPropertyDateValue private; they were previously replaced by
  get_DateLastModified and get_MimeType. All of These methods and
  their helpers are unchanged except for access.
